// src/app/shared/models/enums/button.ts
export enum ButtonContent {
  TextOnly = 'TextOnly',
  IconText = 'IconText',
  IconOnly = 'IconOnly',
}

export enum ButtonSize {
  Small = 'Small',
  Big = 'Big',
}

export enum ButtonOrder {
  Primary = 'Primary',
  Outline = 'Outline',
}

export enum ButtonType {
  Submit = 'Submit',
  Reset = 'Reset',
  Button = 'Button',
}